refactor(prompt): extract role-aware context sections from promptBuilder
promptBuilder.ts (1965 LOC) had absorbed the entire per-role context-builder
family (planner/implementer/reviewer/scm task context, strategy context, and
the buildTaskContext dispatcher) alongside its template-rendering and budget
core. Move that family to src/prompt/contextSections.ts and the shared pure
helpers it needs (formatOptional, compactList, roleContextProfile) to
src/prompt/promptText.ts, giving a clean one-way dependency
(promptText <- contextSections <- promptBuilder) with no import cycle.

Behaviour-preserving: the golden prompt snapshots pass unchanged, confirming
byte-identical output. promptBuilder.ts shrinks to 1629 LOC.
